v3/requirements.txt
streamlit
pandas
pillow
reportlab
opencv-python
numpy
openai
bcrypt
supabase
rapidocr-onnxruntime
